Convert PrivateInfoDialogFragment to Kotlin

+++ app/src/main/java/com/boardgamegeek/ui/model/PrivateInfo.kt
@@ -2,9 +2,9 @@ package com.boardgamegeek.ui.model
 
 data class PrivateInfo(
         val priceCurrency: String? = null,
-        val price: Double = 0.toDouble(),
+        val price: Double = 0.0,
         val currentValueCurrency: String? = null,
-        val currentValue: Double = 0.toDouble(),
+        val currentValue: Double = 0.0,
         val quantity: Int = 1,
         val acquisitionDate: String? = null,
         val acquiredFrom: String? = null,
